refactor(print-format-builder): extract FieldPropertiesPanel
